+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/test/java/org/eclipse/dataspaceconnector/ids/api/multipart/MultipartControllerIntegrationTest.java
@@ -64,8 +64,7 @@ public class MultipartControllerIntegrationTest extends AbstractMultipartControl
         Response response = httpClient.newCall(request).execute();
 
         // verify
-        assertThat(response).isNotNull()
-                .extracting(Response::code).isEqualTo(200);
+        assertThat(response).isNotNull().extracting(Response::code).isEqualTo(200);
 
         List<NamedMultipartContent> content = extractNamedMultipartContent(response);
 
@@ -125,8 +124,7 @@ public class MultipartControllerIntegrationTest extends AbstractMultipartControl
         Response response = httpClient.newCall(request).execute();
 
         // verify
-        assertThat(response).isNotNull()
-                .extracting(Response::code).isEqualTo(200);
+        assertThat(response).isNotNull().extracting(Response::code).isEqualTo(200);
 
         List<NamedMultipartContent> content = extractNamedMultipartContent(response);
 
@@ -195,8 +193,7 @@ public class MultipartControllerIntegrationTest extends AbstractMultipartControl
         Response response = httpClient.newCall(request).execute();
 
         // verify
-        assertThat(response).isNotNull()
-                .extracting(Response::code).isEqualTo(200);
+        assertThat(response).isNotNull().extracting(Response::code).isEqualTo(200);
 
         List<NamedMultipartContent> content = extractNamedMultipartContent(response);